src/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
pyotp==2.9.0
ippanel==2.0.7
psycopg2-binary==2.9.9
pytz==2024.1
django-allauth==0.63.1
oauthlib==3.2.2
requests-oauthlib==2.0.0
python-decouple==3.8
celery==5.4.0
requests==2.31.0
cryptography==42.0.7
django-compressor==4.4
daphne==4.1.2
redis==5.0.8
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-debug-toolbar==4.4.6
django-jalali==6.0.0
python-dateutil==2.9.0.post0
django-mailbox==4.10.1